feat(*): refactor, create dev-test-button class, try to change project and render it
dom.js: split createMain() function to createTasksHeader() and createTasksList(). Use whichActive()
function instead of making it a variable since it's state is always changing. handler.js: move
addTask and removeTask function to Project function factory (become it's method). main.js: Test it
if it can become the controller module, try to change project and render (update) the page if
dev-button is clicked, it works!